InputContainer 코드 작성 & App 코드 리팩토링
