+++ src/App.jsx
@@ -1,38 +1,18 @@
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 
-import Input from './Input';
 import List from './List';
-import { updateRestaurant, addRestaurant } from './actions';
+import InputContainer from './containers/InputContainer';
 
 export default function App() {
-  const dispatch = useDispatch();
-
-  const { restaurants, restaurant } = useSelector((state) => ({
+  const { restaurants } = useSelector((state) => ({
     restaurants: state.restaurants,
-    restaurant: state.restaurant,
   }));
-
-  const handleChangeRestaurant = (e) => {
-    dispatch(
-      updateRestaurant(e.target.name, e.target.value),
-    );
-  };
-
-  const handleClickAddButton = () => {
-    dispatch(
-      addRestaurant(restaurant),
-    );
-  };
 
   return (
     <div>
       <h1>Restaurants</h1>
       <List restaurants={restaurants} />
-      <Input
-        restaurant={restaurant}
-        onChange={handleChangeRestaurant}
-        onClick={handleClickAddButton}
-      />
+      <InputContainer />
     </div>
   );
 }
